Move the rendering cache lock from middleware to render_page method
When a page is renderd in Sitepress, often times there are calls back to
the site within the page that iterates over the site. For example:

```
current_page.children.each do |page|
  = link_to page.data.fetch("title"), page.request_path
```

In a development environment when `Sitepress.site.cache_resource=false`,
the website would have to always rebuild references to these resources,
which is slow. We want this cache to be turned off so that when a developer
changes a page, the change is picked up and can be seen in a browser, but
we want to turn the cache on only while the page is being rendered.

Before this commit, I had this in a middleware `Sitepress::Middleware::RequestCache`
that would manage this development cache, but when using it with a multi-threaded
server like Puma, the value would be changed when another seperate request
was being fullfilled and cause the page to never reload.

Since I'm not that familiar with writing threaded code in Ruby and don't
understand how threaded middleware would impact rails performance, I moved
this concern from middleware into the controller. I figure even if I do
need to thread this in the future, there's a lot less contention for if
I do it at this level.

That said, something feels a bit off about the need for this cache in the
first place. Ideally I could pass off some sort of context to the renderer
that could be called repeatedly without any sort of performance penality.
In a development environment the context would regenerate each time per
rendering call, and in a production environment it would only be generated
once and cached. We'll see where this goes in the future. While this is
better than before, something still feels wrong.