Add support for default pages and extensions
This solves the "root page" problem in Sitepress for the following use cases:

1. `/index.html.haml` would previously resolve to `/index.html`. This is mostly not desirable when running
Sitepress in a server. The more common use case is for that resource to be found via requesting `/`.

2. `/cars/index.html.haml` would similarly resolve to `/cars/index.html`. The desired routing would be that `/cars`.

Changes now need to be made to the static compiler so that `index.html` files get written back out when the
compiler is run for static websites.