Build nodes recursively from source files
Before the nodes would be built from a big list of paths, which would
require a lot of splitting of paths and allocated a ton of arrays.

This approach loops through a directory and adds nodes per object. If the
object is a file, it will add a resource to the node. If its a directory,
it will create a node with the name for that directory and that iterate
through there and ...

This simplifies the code and requires less array allocations