Use Rails for the preview server and compiler
Turns out Rails already has most of the helpers and rendering infrastructure
that I'd otherwise have to roll myself or use something else, like Padrino,
which doesn't have as big of an ecosystem as Rails.

Another benefit of using Rails for the static site compiler is that it makes
it really easy for people to move content sites out of large rails applications
into their own static sites, or to move a Sitepress static site into a dynamic
Rails application.

Another benfit of using Rails is that it is much easier to extend the `sitepress-server`
editing experience. For example, I could use the Rails file watchers to trigger a page reload
via ActionCable when a user saves a file change.

I still am keeping all of the agnostic Sitepress stuff in `sitepress-core`, which
can still be embedded anywhere, but the `sitepress-server` and `sitepress-cli` use
Rails for the development server and compiler.