Add resource rendering post-processing to the Sitepress::Pages controller
This makes it possible to intercept the rendered output of a Sitepress page
for additional processing, say via Nokogiri to change markdown image tags
to use `asset_path` helpers.

For now this can only be accessed from the controller, so its not easy
for users to discover yet. An API should be added in the future that makes
it possible to add rendition processors from the Sitepress configuration.