Bring format, handler, and mime_type into the resource instead of delegating
Turns out we need to be able to modify these independent of the asset depending
on what type of site we're dealing with.

For example, if you want to serve up a Notion site, it only has `blah.md`,
not `blah.html.md`. Thus you'd need to manually set these resources from
the `SourceNodeMapper` when creating a Notion site.