Refactor parsing to allow plugging different parsers into resources
This is the groundwork for making Sitepress work to serve up Notion sites
and parse metadata for other assets like image exif data.

Since notion has CSV files, I baked a CSV parser into the mix so that data
could be extracting from that.

There's still a lot of uncertainty around is this is actually the right
architecture. I think there's a chance this was over-refactored and some
contexts for parsing might be missing (mainly the resource that's being
parsed). We'll see how this pans out in practice, but it could result in a
few changes for this API.