Fix error loading haml templating engine for sites
It would be better if HAML wouldn't load via `ActiveSupport.on_load(:action_view)`,
but it does, so we have to set extensions for Sitepress that way too.

Rails throws up a bunch of deprecation messages in the specs as a result,
but I don't think this affects the code now or in the future.