Bump version from 2.x to 3.x
While the change is small, it drastically changes how Sitepress mounts into a Rails project.

Instead of being an engine that's mounted into Rails, it now integrates directly into the
root applications routes file.

Why the change? I found in practice that content wasn't as seperate from the site that allowed
mounting it as an engine. The biggest problem is I could not set Sitepress as the root page
for a rails project.

This change adds route helpers that makes it much easier to mount Sitepress into rails projects.

The README files are all updated to reflect the changes made from 2.x to 3.x. I really should have
a changelog for this, but I don't, so this will do.